+++ taf/tests/test_git/test_git.py
@@ -36,5 +36,8 @@ def test_is_git_repository_root_non_bare(repository):
 def test_head_commit_sha():
     with tempfile.TemporaryDirectory() as tmpdirname:
         repo = GitRepository(path=tmpdirname)
-        with pytest.raises(GitError, match=f"{tmpdirname} is not a git repository"):
+        with pytest.raises(
+            GitError,
+            match=f"Repo {repo.name}: The path '{repo.path.as_posix()}' is not a Git repository.",
+        ):
             repo.head_commit_sha() is not None